startup chat/requirements.txt
streamlit
python-dotenv
google-generativeai
pathlib